Add files via upload
This is the first upload of the notebook for Case Study 1. Twitter handle is obtained and twitter api is used to search for tweets containing the hashtag #PokemonGo. Resulting data is written to a file called pokemonresults.txt which can be found in the same folder as the iPython notebook.  One search result is then displayed in pretty printed format to the notebook for viewing (and to demonstrate that everything worked).